Minor Update
Added the basics of the Expert NPC class, with more to do later
possibly. Also re-arranged how the Armor section is organized so that
it's easier to use as a reference (all the rules are in one subsection,
and all the armors and shields are in sections according to type).