Crusader Update
Added the Crusader class. The maneuvers and stances might need to be
moved to another section later, but right now it's all in the same very
long entry. Also added a "blank" class entry file so that adding more
classes will be easier in the future. Even though I already have most of
the classes in. Oh well.