Began the Revision
Added the SRD and core Tome source material into some side directories,
with the intent of going through and doing a revision using my latest
set of LaTeX skills. Most stuff will stay unchanged, but I'll attempt to
fill in more chapters this time, combining the SRDpdf source and the
current source. The .rtf and .txt files will be my "record" of what's
been checked and inserted so far.